teams: new Add Team Member button
Per the discussion on #22054, the flow for adding a new team mamber is not intuitive for new Gitea users.

The ideal solution would be to add a new button on the Org > Members index view (see the screenshot mockup in the issue description). However, this would require a refactor of the UX for the flow. The current flow has an implicit context of which team within the org the new member is being added to. From the Members index, there is no implicit context; the flow would have to add a picker for which team the new member should be added to.

So, as a stopgap, this change simply adds a button to the Teams index page that performs the same action as clicking on the title of the team (a behavior that is currently too obscure as indicated in the comments on the issue). This should reduce support burden and serve as a decent temporary measure until the Add Member flow is refactored.